+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/StorageDateVo.java
@@ -8,7 +8,7 @@ import java.sql.Date;
 @Data
 public class StorageDateVo {
     private Date date;
-    private BigInteger daily_sum;
+    private double daily_sum;
     private String storage_name;
     private BigInteger total;
 }
